refactor: change interface
