+++ src/index.ts
@@ -1,5 +1,4 @@
 import Schnorrkel from './schnorrkel'
-export { default as UnsafeSchnorrkel } from './unsafe-schnorrkel'
 
 export { Key, KeyPair, Signature, PublicNonces, Challenge, SignatureOutput, FinalPublicNonce } from './types'
