+++ tests/schnorrkel/verify.test.ts
@@ -16,17 +16,42 @@ describe('testing verify', () => {
 
     const publicNonces = [publicNoncesOne, publicNoncesTwo]
     const publicKeys = [keyPairOne.publicKey, keyPairTwo.publicKey]
-
     const combinedPublicKey = Schnorrkel.getCombinedPublicKey(publicKeys)
 
     const msg = 'test message'
-    const signatureOne = schnorrkelOne.multiSigSign(keyPairOne.privateKey, msg, publicKeys, publicNonces)
-    const signatureTwo = schnorrkelTwo.multiSigSign(keyPairTwo.privateKey, msg, publicKeys, publicNonces)
+    const signatureOne = schnorrkelOne.multiSigSign(keyPairOne.privateKey, msg, combinedPublicKey, publicNonces)
+    const signatureTwo = schnorrkelTwo.multiSigSign(keyPairTwo.privateKey, msg, combinedPublicKey, publicNonces)
 
     const signatures = [signatureOne.signature, signatureTwo.signature]
     const signaturesSummed = Schnorrkel.sumSigs(signatures)
-    const result = Schnorrkel.verify(signaturesSummed, msg, signatureTwo.finalPublicNonce, combinedPublicKey)
+    const result = Schnorrkel.verify(signaturesSummed, msg, signatureTwo.finalPublicNonce, combinedPublicKey.combinedKey)
 
     expect(result).toEqual(true)
+  })
+
+  it('should fail to verify signatures', () => {
+    const schnorrkelOne = new Schnorrkel()
+    const schnorrkelTwo = new Schnorrkel()
+
+    const keyPairOne = generateRandomKeys()
+    const keyPairTwo = generateRandomKeys()
+    const keyPairThree = generateRandomKeys()
+    const publicNoncesOne = schnorrkelOne.generatePublicNonces(keyPairOne.privateKey)
+    const publicNoncesTwo = schnorrkelTwo.generatePublicNonces(keyPairTwo.privateKey)
+
+    const publicNonces = [publicNoncesOne, publicNoncesTwo]
+    const publicKeys = [keyPairOne.publicKey, keyPairTwo.publicKey]
+    const combinedPublicKey = Schnorrkel.getCombinedPublicKey(publicKeys)
+    const combinedPublicKeyTwo = Schnorrkel.getCombinedPublicKey([keyPairOne.publicKey, keyPairThree.publicKey])
+
+    const msg = 'test message'
+    const signatureOne = schnorrkelOne.multiSigSign(keyPairOne.privateKey, msg, combinedPublicKey, publicNonces)
+    const signatureTwo = schnorrkelTwo.multiSigSign(keyPairTwo.privateKey, msg, combinedPublicKey, publicNonces)
+
+    const signatures = [signatureOne.signature, signatureTwo.signature]
+    const signaturesSummed = Schnorrkel.sumSigs(signatures)
+    const result = Schnorrkel.verify(signaturesSummed, msg, signatureTwo.finalPublicNonce, combinedPublicKeyTwo.combinedKey)
+
+    expect(result).toEqual(false)
   })
 })
